+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Deprecated
+
+- CLI flag `--wasm` and `--contract-path` are deprecated in favor of `--contract`
+- CLI flag `--snapshot` is deprecated in favor of `--network-snapshot`
+
 ## [0.1.0] - 2026-02-19
 
 ### Added